refactor: moved resource routes to new file + fixes in the resources usage example
